feat : 파티 만료시간 컴포넌트로 분리
- 파티 만료시간 컴포넌트(RemainingTime)로 분리
- 카드, 카드 상세보기에 RemainingTime 적용
